validate ccu firmware version

+++ download_fw_ccu.sh
@@ -1,3 +1,4 @@
+#!/bin/sh
 TARGETDIR=/usr/local/etc/config/firmware
 TMPDIR=/tmp/fwtmp
    
@@ -5,6 +6,10 @@ TMPDIR=/tmp/fwtmp
 if [[ "$PARAM" == "-o" ]]; then echo "Overwriting existing contents!";fi
 if [[ "$PARAM" == "-c" ]]; then echo "Clearing all firmware files"; rm -rf ${TARGETDIR}/*; fi
 
+ccu_model=`grep VERSION /boot/VERSION   | awk -F'[=.]' {'print $2'}`
+ccu_version=`grep VERSION /boot/VERSION | awk -F'[=.]' {'print $3'}`
+ccu_build=`grep VERSION /boot/VERSION   | awk -F'[=.]' {'print $4'}`
+echo "Running CCU3 Firmware $ccu_model.$ccu_version.$ccu_build"
 echo "Getting firmware list"
 output=`curl -s 'https://update.homematic.com/firmware/api/firmware/search/DEVICE' | sed s/'homematic.com.setDeviceFirmwareVersions('/''/ | sed s/');'/''/ |  jq -r '.[] | "\(.type)"'`
 
@@ -25,15 +30,33 @@ for row in ${output}; do
     tar -xzf $f
     rm $f
     id=`grep TypeCode info | awk -F= {'print $2'}|xargs`
-    echo "${id}]"
-    if [ ! -d ${TARGETDIR}/$id ]; then
-      mkdir -p ${TARGETDIR}/$id
-      mv * ${TARGETDIR}/$id/
-    else
-      if [[ "$PARAM" == "-o" ]]; then
-        rm -rf ${TARGETDIR}/$id/*
-        mv * ${TARGETDIR}/$id/
+    fw_model=`grep CCU3FirmwareVersionMin info | awk -F'[=.]' {'print $2'}`
+    fw_version=`grep CCU3FirmwareVersionMin info | awk -F'[=.]' {'print $3'}`
+    fw_build=`grep CCU3FirmwareVersionMin info | awk -F'[=.]' {'print $4'}|xargs`
+    echo -n "${id}] FW min. Version: $fw_model.$fw_version.$fw_build "
+    fw_ok=0
+    if [ "$fw_model" == "$ccu_model" ]; then 
+      if [ $fw_version -lt $ccu_version ]; then
+        fw_ok=1
+      elif [ $fw_version -eq $ccu_version ]; then 
+        if [ $fw_build -le $ccu_build ]; then                                                                 
+         fw_ok=1                                                           
+        fi
       fi
+    fi
+    if [ $fw_ok -eq 1 ]; then
+      echo "OK"
+      if [ ! -d ${TARGETDIR}/$id ]; then       
+        mkdir -p ${TARGETDIR}/$id       
+        mv * ${TARGETDIR}/$id/          
+      else                              
+        if [[ "$PARAM" == "-o" ]]; then 
+          rm -rf ${TARGETDIR}/$id/*     
+          mv * ${TARGETDIR}/$id/        
+        fi                             
+      fi
+    else
+      echo "NOT OK"
     fi
   done
 done
